backend/requirements.txt
azure-functions==1.18.0
azure-storage-blob==12.19.0
Pillow==10.4.0
numpy==1.26.4
matplotlib==3.9.2
requests>=2.28.0